Fix rounding

+++ ggc-core/src/ggc/DerivativeProduct.java
@@ -36,7 +36,7 @@ public class DerivativeProduct extends Product {
 
     @Override
     public String toString() {
-        return getId() + "|" + getMaxPrice() + "|" + getStock() + "|" + _multiplier + "|" + _recipe.toString();
+        return getId() + "|" + (int)getMaxPrice() + "|" + getStock() + "|" + _multiplier + "|" + _recipe.toString();
     }
 
 }
